requirements.txt
appdirs==1.4.4
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.9.3
black==21.11b1
bs4==0.0.1
certifi==2020.12.5
cffi==1.15.0
cfgv==3.3.1
chardet==4.0.0
click==8.0.3
cryptography==35.0.0
defusedxml==0.7.1
distlib==0.3.1
dj-rest-auth==2.1.11
Django==4.0.1
django-admin-list-filter-dropdown==1.0.3
django-allauth==0.47.0
django-countries==7.2.1
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==5.0.0
fake-headers==1.0.2
filelock==3.0.12
GDAL==3.3.3
h11==0.12.0
html5lib==1.1
identify==2.4.0
idna==2.10
lxml==4.7.1
Markdown==3.3.4
mypy-extensions==0.4.3
mysqlclient==2.1.0
nodeenv==1.6.0
numpy==1.21.4
oauthlib==3.1.1
outcome==1.1.0
pathspec==0.9.0
pbr==5.5.1
platformdirs==2.4.0
pre-commit==2.16.0
protobuf==3.17.3
pycparser==2.20
pydub==0.25.1
PyJWT==2.3.0
pyOpenSSL==21.0.0
python3-openid==3.2.0
pytz==2021.1
PyYAML==6.0
regex==2021.11.10
requests==2.25.1
requests-oauthlib==1.3.0
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2
sqlparse==0.4.1
stevedore==3.3.0
timezonefinder==5.2.0
toml==0.10.2
tomli==1.2.2
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.1
Unipath==1.1
urllib3==1.26.3
virtualenv==20.4.2
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
webencodings==0.5.1
wsproto==1.0.0
